requirements.txt
aiohttp>=3.8.0
PyYAML>=5.1 